Fixes possible deadlock if a signal is sent to application before io_service.
When core.close() happens, the tap adapter is closed (file descriptor) but
it is possible that in another thread, the do_read_tap() handler is executed and
a new async_read() is registered. As the file descriptor is closed, we have an
endless loop with read error."

It has been detected using valgrind and hitting CTRL-C at the wrong moment
(i.e. in the middle of FSCP initialization) and we get endless loop of "Failed
to read on tap0", 100% CPU and a "kill -KILL" is mandatory to exit the
application.